using smaller variables in AppStorage, packed structs

+++ contracts/interfaces/rates/ILinearKinkedRateModel.sol
@@ -2,8 +2,8 @@
 pragma solidity 0.8.4;
 
 interface ILinearKinkedRateModel {
-    function baseRate() external view returns(uint256);
-    function optimalUtilRate() external view returns(uint256);
-    function slope1() external view returns(uint256);
-    function slope2() external view returns(uint256);
+    function baseRate() external view returns(uint64);
+    function optimalUtilRate() external view returns(uint64);
+    function slope1() external view returns(uint64);
+    function slope2() external view returns(uint64);
 }
