+++ contracts/interfaces/rates/ILogDerivativeRateModel.sol
@@ -2,7 +2,7 @@
 pragma solidity 0.8.4;
 
 interface ILogDerivativeRateModel {
-    function baseRate() external view returns(uint256);
-    function factor() external view returns(uint256);
-    function maxApy() external view returns(uint256);
+    function baseRate() external view returns(uint64);
+    function factor() external view returns(uint80);
+    function maxApy() external view returns(uint80);
 }
